* Renamed a file and introduced a mixin for actor inventory calls.
