+++ src/module/item/mixins/item-activation.js
@@ -1,5 +1,3 @@
-import { ActorItemHelper, getChildItems, getItemContainer, setItemContainer } from "../../actor/actor-inventory.js"
-
 export const ItemActivationMixin = (superclass) => class extends superclass {
 
     hasUses() {
