+++ src/sfrpg.js
@@ -11,7 +11,7 @@ import { preloadHandlebarsTemplates } from "./module/templates.js";
 import { registerSystemSettings } from "./module/settings.js";
 import { measureDistances, getBarAttribute, handleItemDropCanvas } from "./module/canvas.js";
 import { ActorSFRPG } from "./module/actor/actor.js";
-import { initializeRemoteInventory, ActorItemHelper } from "./module/actor/actor-inventory.js";
+import { computeCompoundBulkForItem, initializeRemoteInventory, ActorItemHelper } from "./module/actor/actor-inventory-utils.js";
 import { ActorSheetSFRPGCharacter } from "./module/actor/sheet/character.js";
 import { ActorSheetSFRPGDrone } from "./module/actor/sheet/drone.js";
 import { ActorSheetSFRPGHazard } from "./module/actor/sheet/hazard.js";
@@ -31,7 +31,6 @@ import { generateUUID } from "./module/utilities.js";
 import migrateWorld from './module/migration.js';
 import CounterManagement from "./module/classes/counter-management.js";
 import templateOverrides from "./module/template-overrides.js";
-import { computeCompoundBulkForItem } from "./module/actor/actor-inventory.js";
 import { RPC } from "./module/rpc.js";
 import { DiceSFRPG } from './module/dice.js';
 import { ActorSheetFlags } from './module/apps/actor-flags.js';
